Extract text from pdf in reading order using pdftotext
I used the following command, running from the parent folder of 'step1':

  pdftotext ipcc-countries.pdf step1/ipcc-countries.txt

I also experimented with the option -layout, which preserves the layout
of the file, keeping values in the left column and values in the right
column on a same line. The result is closer to the target format without
this option.