Remove extra empty line from end of xsltproc output
In some cases, I had output commas to create an empty record
and present this empty line as CSV.

I decided to remove the line completely from the shell,
using 'head -n -1'.

This preserves the XSLT transformation code as expected,
and I documented the motivation for the extra step with
a comment in each shell script involved.

I checked that the change introduced no other change
than removing the empty line/record from the end of
the file, and updated the Checked field in meta.txt
to today's date.