Prepare to add data sources for country flags, from Wikipedia
In these data sources, the data.csv file will list country names
and relative paths for flag images, which associated to different
base Uniform Resource Locators (URL), give access to the description
page of the image on Wikipedia or Wikimedia Commons.

From the description pages, the URL of the actual images can be found,
together with a description of the image and its conditions of use.

Note: although this URL can be interpolated from the URL
of the thumbnail referenced in the src attribute of the image,
by removing 'thumb/'
and cutting the URL after the first occurrence of .svg,
I will download each HTML page as a separate source
to keep a full track record of the origin
and the history of the image,
including conditions of use.