Start to add parsing step
I plan to use two different style sheets to extract data from country
pages and from flag pages. The name of the style sheet differs only
while in step4 folder in this project; it is provided as parameter
to parse() and the file is then copied to parse-data.xsl in step3
folder of the outside source.