Use method="text" instead of mode="text" in xsl:output
This is the same confusion as before, already fixed in master branch.
I had to keep the extra step (head -n -1) to remove the last line here,
because it removes an empty record generated by the transformation
(which is easier to write this way). Although I could have left the
empty record in the file (no harm done), I decided to remove it to
keep the file data.csv unchanged with this commit.