Add description to generated meta.txt
At this point, I am not convinced of the utility of the first paragraph
of the page, extracted for the description. Also, if I extracted this
paragraph by hand, I would remove the least relevant parts (parentheses
about pronunciation and notes in square brackets), as is done in the
page preview in results of duckduckgo search engine, and I would wrap
the text at relevant positions to fit within a width of less than
80 characters when the file is viewed in a terminal.

Since a copy of the page will be included anyway, I will rather remove
the unpolished extract than complete the unneeded feature.