Add start of step 3: create separate external sources for each entity
I have implemented all steps up to the parsing of XML data
and the extraction of parts of the data in CSV.

Only one type of value is supported in claims so far.
I have included code to detect unsupported types,
and will use it to complete support for different types of values.