Add first version of the E/R diagram generated by erd
I am only partly done documenting the relationships between tables,
which enlightens the database as it reflects the flow of data from
one step to the next. I am currently using the definitions of the
views to describe their relationships to the underlying tables.
I will then describe the relationships that correspond to the
redundant information cross-checked in step 2.