Export the full database in SQL format to database/ folder
The complete database will be used in queries to produce
data needed for the visualizations. While the selection
of views in CSV format could be amended to contain every
detail needed in a visualization, I feel like this is a
limitation to the expressiveness of visualizations, which
should be free to pick both from the main information in
these views completed with details from the whole database.