Use SVG sprite images
The external SVG file must be in the same folder
for the reference to work, at least from a local file, in Firefox 37.
I thus had to make a copy of flags.svg in step3 folder.

The view displayed for each flag is not correct at this point,
only a part of the top left corner is visible.