Remove TODO: the shorthand notation of paths is the only one allowed
I assumed that there was a long version of the commands, using the
full name, e.g. 'moveto', which could then be shortened to 'M' or 'm'.
Actually, the only representation allowed for moveto command in the
grammar is 'M' or 'm'.

Reference:

8.3.9 The grammar for path data
in 8. Paths
in SVG 1.1 (Second Edition) – 16 August 2011
http://www.w3.org/TR/SVG/paths.html#PathDataBNF