Avoid matching header record by mistake when no flag is found
This issue occurs with the updated csvgrep command, which apparently
never matches any flag. In this case, the header record remains, where
no record is expected.

EDIT: the issue is not related to the updated csvgrep, but to a change
in wikidata API apparently, which returns an error stating that 'xml'
is not a supported media type. I will investigate.