Do not add 'https:' protocol in duplicate to wikimedia URL for SVG files
While the URL for the original picture in SVG format used to start
with '//' in HTML pages for flags on Wikimedia, i.e. they used a protocol-
relative URL in which the protocol of the link is identical to the one of
the page, the URL now includes the 'https' protocol already.