Update country information and country flags
The whole process starts from the scripts in
wikidata.org/2015-api-query-backlinks-to-iso-alpha3/,
which query the list of entities with ISO3 country codes
from wikidata.org.

There is then a separate process to retrieve data for each
country from wikidata.org and extract the reference of the flag.

The flag page is then downloaded from wikimedia, as well as the
referenced SVG file (original image).