Add 'https' protocol to URL for SVG picture, only when missing
I noticed two errors in the log of the latest update:

  curl: (3) <url> malformed

They correspond to URLs still in the former format,
starting with '//' (protocol-relative URL).

I will apply this fix just to these two country flags
to avoid rerunning the whole process at this point;
the fix will be generalized with the next update instead.